feat: シーシャ店舗APIフックを削除
